feature: guess animal points increasing implemented

+++ core/data/profiles/api/src/main/java/ru/heatrk/languageapp/core/profiles/api/domain/Profile.kt
@@ -5,6 +5,6 @@ data class Profile(
     val email: String?,
     val firstName: String?,
     val lastName: String?,
-    val totalScore: Int,
+    val totalScore: Float,
     val avatarUrl: String?,
 )
